+++ frontend/app_flowy/lib/plugins/grid/presentation/widgets/toolbar/grid_toolbar.dart
@@ -1,14 +1,10 @@
-import 'package:app_flowy/plugins/grid/application/setting/setting_bloc.dart';
-import 'package:flowy_infra/image.dart';
-import 'package:flowy_infra_ui/flowy_infra_ui.dart';
-import 'package:flowy_infra_ui/style_widget/extension.dart';
-import 'package:flowy_infra_ui/style_widget/icon_button.dart';
 import 'package:flutter/material.dart';
 
 import '../../../application/field/field_controller.dart';
 import '../../layout/sizes.dart';
-import 'grid_property.dart';
+import 'filter_button.dart';
 import 'grid_setting.dart';
+import 'setting_button.dart';
 
 class GridToolbarContext {
   final String gridId;
@@ -34,68 +30,11 @@ class GridToolbar extends StatelessWidget {
       child: Row(
         children: [
           SizedBox(width: GridSize.leadingHeaderPadding),
-          _SettingButton(settingContext: settingContext),
+          SettingButton(settingContext: settingContext),
           const Spacer(),
+          FilterButton(),
         ],
       ),
-    );
-  }
-}
-
-class _SettingButton extends StatelessWidget {
-  final GridSettingContext settingContext;
-  const _SettingButton({required this.settingContext, Key? key})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return AppFlowyPopover(
-      constraints: BoxConstraints.loose(const Size(260, 400)),
-      offset: const Offset(0, 10),
-      margin: const EdgeInsets.all(6),
-      child: FlowyIconButton(
-        width: 22,
-        icon: svgWidget(
-          "grid/setting/setting",
-          color: Theme.of(context).colorScheme.onSurface,
-        ).padding(horizontal: 3, vertical: 3),
-      ),
-      popupBuilder: (BuildContext context) {
-        return _GridSettingListPopover(settingContext: settingContext);
-      },
-    );
-  }
-}
-
-class _GridSettingListPopover extends StatefulWidget {
-  final GridSettingContext settingContext;
-
-  const _GridSettingListPopover({Key? key, required this.settingContext})
-      : super(key: key);
-
-  @override
-  State<StatefulWidget> createState() => _GridSettingListPopoverState();
-}
-
-class _GridSettingListPopoverState extends State<_GridSettingListPopover> {
-  GridSettingAction? _action;
-
-  @override
-  Widget build(BuildContext context) {
-    if (_action == GridSettingAction.properties) {
-      return GridPropertyList(
-        gridId: widget.settingContext.gridId,
-        fieldController: widget.settingContext.fieldController,
-      );
-    }
-
-    return GridSettingList(
-      settingContext: widget.settingContext,
-      onAction: (action, settingContext) {
-        setState(() {
-          _action = action;
-        });
-      },
     );
   }
 }
